* Cash weapons now showing properly in character avatars (character select page, 3rd person views from map, smegas, messenger, etc.)  * Slightly rewrote visible/masked equips logic in CommonPackets.writeAvatar to use less conditionals.  * Fixed map crash when a player set a mastery level for a ranged weapon and they knock-backed a mob.  * Fixed self crash when a player has a cash equip. Rings no longer crash, but ring effects still aren't implemented yet.  * Added !maxequips, !maxskills, and !maxall commands.  * !spawn command updated to support specifying an amount of mobs to spawn.  * LittleEndianWriter.writePaddedAsciiString and writeNullTerminatedString will no longer throw NullPointerException.

+++ trunk/src/argonms/common/character/inventory/InventorySlot.java
@@ -35,7 +35,7 @@ public abstract class InventorySlot implements Comparable<InventorySlot>, Clonea
 	private long expire;
 	private long uid;
 	private String owner;
-	private byte flag;
+	private short flag;
 
 	protected InventorySlot(int itemid) {
 		this.id = itemid;
@@ -74,11 +74,11 @@ public abstract class InventorySlot implements Comparable<InventorySlot>, Clonea
 		this.owner = owner;
 	}
 
-	public byte getFlag() {
+	public short getFlag() {
 		return flag;
 	}
 
-	public void setFlag(byte flag) {
+	public void setFlag(short flag) {
 		this.flag = flag;
 	}
